requirements.txt
paddlepaddle==2.5.1
paddleocr
Pillow
opencv-python
openai
numpy
colormath
easyocr